Added Environment Upkeeper widget
DO NOT EVER DISABLE THIS WIDGET
