--- PackageTemplate/my_lib.py
@@ -1,4 +1,0 @@
-from Py4GWCoreLib import *
-
-def my_library_function():
-    return "This is a function from my_lib.py"
