+++ Py4GW.ini
@@ -26,8 +26,8 @@ enabled = True
 subcategory = Triggers
 
 [Widget Manager]
-x = 353
-y = 5
+x = 434
+y = 4
 enable_all = True
 collapsed = True
 
@@ -55,6 +55,11 @@ enabled = True
 category = Coding
 subcategory = Info
 enabled = False
+
+[Environment Upkeeper]
+category = Coding
+subcategory = Environment
+enabled = True
 
 [Titles]
 category = Gameplay
