+++ imgui.ini
@@ -16,20 +16,20 @@ Collapsed=0
 [Window][Py4GW##compactPy4GWconsole]
 Pos=60,60
 Size=128,178
-Collapsed=1
+Collapsed=0
 
 [Window][Widgets]
-Pos=353,5
-Size=105,164
+Pos=434,4
+Size=121,224
 Collapsed=1
 
 [Window][Loot Manager]
-Pos=250,43
-Size=184,339
-Collapsed=1
+Pos=60,60
+Size=184,342
+Collapsed=0
 
 [Window][Switch Character]
-Pos=443,43
+Pos=60,60
 Size=320,254
 Collapsed=0
 
@@ -51,5 +51,10 @@ Collapsed=1
 [Window][##MissionMapOverlay]
 Pos=1137,30
 Size=533,504
+Collapsed=0
+
+[Window][##invisible_windowMega Zoom]
+Pos=1137,507
+Size=224,25
 Collapsed=0
